Warn about arbitrary command execution in reusable workflow (#8)
Add a note to this project's reusable workflow that some of it inputs
can be used of arbitrary command execution. Given this file isn't really
intended for reuse outside of this project, the assumption is that it
will only reused by someone that finds the file, hence putting the note
in the file should suffice.

Additionally, this hardens the reusable workflow slightly by removing
two instances where workflow expressions unnecessarily allow arbitrary
command execution.