Add monitor for `dorny/paths-filter` at v3 (#12)
Create a new monitor for the dorny/paths-filter action. The monitor is
for the latest major version. The `files:` are based on the files that
exist in the project's `dist/` directory. The build-cmd is based on the
list of commands the project declares in `package.json`. The install-cmd
is chosen based on the fact that the project has a `package-lock.json`
file. The node-version is based on the `"engines"` value in the project
manifest. Other than that, everything is generated by the `new.sh`
script.

All this is based on the state of the repository at commit:
de90cc6fb38fc0963ad72b210f1f284cd68cea36