Add monitor for `actions/attest-build-provenance` at v1 (#20)
Create a new monitor for the actions/attest-build-provenance action.

The monitor is for the latest major version. The files: are based on the files that exist in the project's dist/ directory. The build-cmd is based on the list of commands the project declares in package.json. The install-cmd is chosen based on the fact that the project has a package-lock.json file. The node-version-file is based on the node-version-file value used in the project's own workflows.

All this is based on the state of the repository at commit: 8f1fc17a59ea731736cdf0e179a0dcf1f6270453