Update docs and simplify workflows (#45)
Remove the use of the `<VERSION>` as the job id or name since workflows
cannot contain jobs for multiple version. This stems from the fact that
the badges used in the README are workflow based rather than job-based,
hence we need separate jobs for different versions of the same action to
be able to display their reproducibility status independently.